Skip audit logging for inline-renderable Content-Types
Refine the audit-log gate so we don't record session_file_download
events for files the browser will render inline (PDF, images, text,
audio, video, etc.). The user perceives those as views, not
downloads, and the audit log was misleading by including them.

Factor the now ~30-line audit-policy decision into a single helper
module_context::shouldAuditFileDownload(request, filePath) so both
handlers share one place to evolve the rule. The helper returns
false when:

  1. URL carries ?show=1 (server-initiated preview, or GWT-tagged
     view-style navigation).
  2. Sec-Fetch-Dest indicates a sub-resource fetch (style/script/
     image/etc.).
  3. The file's Content-Type starts with text/, image/, audio/, or
     video/, or matches application/pdf, application/json,
     application/xml, application/javascript, or
     application/xhtml+xml.

Otherwise the request is treated as a download and audited.

Trade-offs: a real "right-click -> Save As" on a rendered PDF won't
be audited (server can't observe that), and a renamed
secret.zip -> secret.txt would skip the gate (text/plain). Both are
acceptable trade-offs for not over-logging the much more common
inline-view case.

Addresses rstudio-pro#11010.
