+++ src/cpp/session/include/session/SessionModuleContext.hpp
@@ -88,12 +88,22 @@ enum PackageCompatStatus
    COMPAT_UNKNOWN = 4
 };
     
-// paths 
+// paths
 core::FilePath userHomePath();
 std::string createAliasedPath(const core::FileInfo& fileInfo);
 std::string createAliasedPath(const core::FilePath& path);
 std::string createFileUrl(const core::FilePath& path);
 core::FilePath resolveAliasedPath(const std::string& aliasedPath);
+
+// Returns true if a /files/ or /file_show request for filePath should
+// produce a session_file_download audit event. Returns false when the
+// request is a server-initiated preview/show (URL has ?show=1), an HTML
+// sub-resource fetch (Sec-Fetch-Dest is style/script/image/etc.), or
+// when the response will be served as a browser-renderable Content-Type
+// (text/*, image/*, audio/*, video/*, application/pdf, etc.) and the
+// bytes will be viewed inline rather than saved.
+bool shouldAuditFileDownload(const core::http::Request& request,
+                             const core::FilePath& filePath);
 core::FilePath userScratchPath();
 core::FilePath userUploadedFilesScratchPath();
 core::FilePath scopedScratchPath();
